requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1.1
django-colorfield==0.11.0
django-import-export==4.1.1
django-quill-editor==0.1.42
django_csp==3.8
djangorestframework==3.15.2
Faker==30.1.0
pillow==11.0.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
